backend/requirements.txt
# Core Flask
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-engineio==4.9.0
eventlet==0.35.2

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
joblib==1.3.2

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# Input Validation
marshmallow==3.20.1
email-validator==2.1.0

# API Documentation
flask-swagger-ui==4.11.1
flasgger==0.9.7.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production Server
gunicorn==21.2.0

# Phase 6: Premium Features
reportlab==4.0.7
requests==2.31.0
schedule==1.2.0


